Moved pathutils to pkg
